pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "podclaw"
version = "0.1.0"
description = "One-prompt podcast production — generate full episodes from a single topic."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "v0id_injector"},
]
keywords = ["podcast", "tts", "elevenlabs", "video", "ai", "audio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "requests>=2.28",
    "Pillow>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
podclaw = "podclaw.cli:main"

[project.urls]
Homepage = "https://github.com/v0id-injector/podclaw"
Repository = "https://github.com/v0id-injector/podclaw"
Issues = "https://github.com/v0id-injector/podclaw/issues"

[tool.setuptools.packages.find]
include = ["podclaw*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
